okhttp3:okhttp and okhttp3:okhttp-sse artifacts were removed
