+++ lib/build.gradle
@@ -49,11 +49,6 @@ dependencies {
     // https://mvnrepository.com/artifact/com.google.code.gson/gson
     implementation 'com.google.code.gson:gson:2.9.1'
 
-    // https://square.github.io/okhttp
-    implementation 'com.squareup.okhttp3:okhttp:4.11.0'
-    // https://mvnrepository.com/artifact/com.squareup.okhttp3/okhttp-sse
-    implementation 'com.squareup.okhttp3:okhttp-sse:4.11.0'
-
     // Adds getter, setter and builder boilerplate
     // https://projectlombok.org/
     // https://mvnrepository.com/artifact/org.projectlombok/lombok
